Trying to get timeouts working in Informix plus fixed Tickets #48-49

+++ trunk/src/main/java/nz/co/gregs/dbvolution/datatypes/DBNumber.java
@@ -157,7 +157,7 @@ public class DBNumber extends QueryableDatatype implements NumberResult {
 		if (newLiteralValue == null) {
 			super.setLiteralValue(null);
 		} else {
-			super.setLiteralValue(newLiteralValue);
+			super.setLiteralValue(newLiteralValue.doubleValue());
 		}
 	}
